Broke the import module out of main.py
