+++ import_module.py
@@ -1,0 +1,21 @@
+import importlib
+import imp
+
+
+def import_module(name):
+    if name[-3:] == ".py":
+        # assume we're working with a path
+        try:
+            return imp.load_source("function", name)
+        except IOError:
+            print("ERROR: Could not find file: {}".format(name))
+            print("Exiting")
+            sys.exit()
+    else:
+        # assume we're working with a module
+        try:
+            return importlib.import_module(name)
+        except ImportError:
+            print("ERROR: Could not find module: {}".format(name))
+            print("Exiting")
+            sys.exit()
